backend/python/requirements.txt
flask==2.3.2
flask-cors==3.0.10
google-cloud-firestore==2.11.0
networkx==3.2
python-dotenv==1.0.0
requests==2.31.0